LPS-66938 Move layout set prototype related data handlers to the proper package
